requirements.txt
# Web framework
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20

# Database
sqlalchemy[asyncio]==2.0.40
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.15.2

# Celery
celery[redis]==5.4.0
redis==5.3.0

# Encryption
cryptography==44.0.3

# Cloud SDKs
boto3==1.38.5
google-cloud-bigquery==3.31.0
azure-identity==1.21.0
azure-mgmt-costmanagement==4.0.1
azure-mgmt-cognitiveservices>=14.0.0
azure-mgmt-resource>=23.0.0
azure-mgmt-subscription>=3.1.0
PyJWT>=2.8.0

# Utilities
python-dateutil==2.9.0
httpx==0.28.1
# Keep pydantic + pydantic-core versions in sync (mismatch breaks FastAPI import).
pydantic[email]==2.11.2
pydantic-core==2.33.1
pydantic-settings==2.9.1
openpyxl==3.1.5
